Copy TOTP to clipboard after passkey authentication in Android (#2305)
